update serializer; update tests
