+++ setup.py
@@ -66,7 +66,7 @@ setup(
     ],
     tests_require=['flake8', 'scikit-image'],
     extras_require={
-        'all': ['pillow', 'scipy', 'h5py', 'lmdb>=0.92', 'matplotlib', 'scikit-learn'],
+        'all': ['scipy', 'h5py', 'lmdb>=0.92', 'matplotlib', 'scikit-learn'],
         'all: "linux" in sys_platform': ['python-prctl'],
         'all: python_version < "3.0"': ['tornado'],
     },
