+++ tensorpack/dataflow/serialize_test.py
