+++ tensorpack/models/__init__.py
@@ -44,5 +44,7 @@ for _, module_name, _ in iter_modules(
         continue
     if module_name.startswith('_'):
         continue
+    if "_test" in module_name:
+        continue
     if module_name not in _SKIP:
         _global_import(module_name)
