+++ tensorpack/utils/serialize.py
@@ -2,13 +2,10 @@
 # File: serialize.py
 
 import os
-import sys
 
+import pickle
 import msgpack
 import msgpack_numpy
-
-from . import logger
-from .develop import create_dummy_func
 
 msgpack_numpy.patch()
 assert msgpack.version >= (0, 5, 2)
@@ -19,83 +16,90 @@ __all__ = ['loads', 'dumps']
 MAX_MSGPACK_LEN = 1000000000
 
 
-def dumps_msgpack(obj):
-    """
-    Serialize an object.
+class MsgpackSerializer(object):
 
-    Returns:
-        Implementation-dependent bytes-like object.
-    """
-    return msgpack.dumps(obj, use_bin_type=True)
+    @staticmethod
+    def dumps(obj):
+        """
+        Serialize an object.
 
+        Returns:
+            Implementation-dependent bytes-like object.
+        """
+        return msgpack.dumps(obj, use_bin_type=True)
 
-def loads_msgpack(buf):
-    """
-    Args:
-        buf: the output of `dumps`.
-    """
-    # Since 0.6, the default max size was set to 1MB.
-    # We change it to approximately 1G.
-    return msgpack.loads(buf, raw=False,
-                         max_bin_len=MAX_MSGPACK_LEN,
-                         max_array_len=MAX_MSGPACK_LEN,
-                         max_map_len=MAX_MSGPACK_LEN,
-                         max_str_len=MAX_MSGPACK_LEN)
-
-
-def dumps_pyarrow(obj):
-    """
-    Serialize an object.
-
-    Returns:
-        Implementation-dependent bytes-like object.
-        May not be compatible across different versions of pyarrow.
-    """
-    return pa.serialize(obj).to_buffer()
+    @staticmethod
+    def loads(buf):
+        """
+        Args:
+            buf: the output of `dumps`.
+        """
+        # Since 0.6, the default max size was set to 1MB.
+        # We change it to approximately 1G.
+        return msgpack.loads(buf, raw=False,
+                             max_bin_len=MAX_MSGPACK_LEN,
+                             max_array_len=MAX_MSGPACK_LEN,
+                             max_map_len=MAX_MSGPACK_LEN,
+                             max_str_len=MAX_MSGPACK_LEN)
 
 
-def loads_pyarrow(buf):
-    """
-    Args:
-        buf: the output of `dumps`.
-    """
-    return pa.deserialize(buf)
+class PyarrowSerializer(object):
+    @staticmethod
+    def dumps(obj):
+        """
+        Serialize an object.
 
-
-# import pyarrow has a lot of side effect:
-# https://github.com/apache/arrow/pull/2329
-# https://groups.google.com/a/tensorflow.org/forum/#!topic/developers/TMqRaT-H2bI
-# So we use msgpack as default.
-if os.environ.get('TENSORPACK_SERIALIZE', 'msgpack') == 'pyarrow':
-    try:
+        Returns:
+            Implementation-dependent bytes-like object.
+            May not be compatible across different versions of pyarrow.
+        """
         import pyarrow as pa
-    except ImportError:
-        loads_pyarrow = create_dummy_func('loads_pyarrow', ['pyarrow'])  # noqa
-        dumps_pyarrow = create_dummy_func('dumps_pyarrow', ['pyarrow'])  # noqa
+        return pa.serialize(obj).to_buffer()
 
-    if 'horovod' in sys.modules:
-        logger.warn("Horovod and pyarrow may have symbol conflicts. "
-                    "Uninstall pyarrow and use msgpack instead.")
-    loads = loads_pyarrow
-    dumps = dumps_pyarrow
+    @staticmethod
+    def dumps_bytes(obj):
+        """
+        Returns:
+            bytes
+        """
+        return PyarrowSerializer.dumps(obj).to_pybytes()
+
+    @staticmethod
+    def loads(buf):
+        """
+        Args:
+            buf: the output of `dumps` or `dumps_bytes`.
+        """
+        import pyarrow as pa
+        return pa.deserialize(buf)
+
+
+class PickleSerializer(object):
+    @staticmethod
+    def dumps(obj):
+        """
+        Returns:
+            bytes
+        """
+        return pickle.dumps(obj, protocol=-1)
+
+    @staticmethod
+    def loads(buf):
+        """
+        Args:
+            bytes
+        """
+        return pickle.loads(buf)
+
+
+_DEFAULT_S = os.environ.get('TENSORPACK_SERIALIZE', 'msgpack')
+
+if _DEFAULT_S == "pyarrow":
+    dumps = PyarrowSerializer.dumps_bytes
+    loads = PyarrowSerializer.loads
+elif _DEFAULT_S == "pickle":
+    dumps = PickleSerializer.dumps
+    loads = PickleSerializer.loads
 else:
-    loads = loads_msgpack
-    dumps = dumps_msgpack
-
-
-class NonPicklableWrapper(object):
-    """
-    TODO
-
-    https://github.com/joblib/joblib/blob/master/joblib/externals/loky/cloudpickle_wrapper.py
-    """
-    def __init__(self, obj):
-        self._obj = obj
-
-    def __reduce__(self):
-        import dill
-        s = dill.dumps(self._obj)
-        return dill.loads, (s, )
-
-    def __call__(self, *args, **kwargs):
-        return self._obj(*args, **kwargs)
+    dumps = MsgpackSerializer.dumps
+    loads = MsgpackSerializer.loads
